refactor(blockstore/engine): inline transfer-request constructors and merge sync_entry.go
The three single-purpose constructors NewDownloadRequest, NewPrefetchRequest,
and NewBlockUploadRequest were thin struct-literal wrappers with no
validation or normalization. Each had one production callsite (or zero, in
the upload case); inline them and use named struct literals directly.

The TransferRequest type and BlockKey method are co-located with TransferType
in types.go; sync_entry.go (45 lines) becomes empty and is removed.

Tests that just verified the constructor field assignments are tautological
once the wrappers are gone; preserve the behavior tests (BlockKey,
TransferType.String) in types_test.go.

Pure refactor. No behavior change.
